requirements.txt
cement
httplib2
python-Levenshtein
google-api-python-client
pyxdg
python-dateutil
